Fix storing floats in Redis
Redis will only accept floats with a point decimal sign and not a comma decimal sign. The sign is determined by the system's locale setting (LC_NUMERIC) and if set to one that prefers a comma will cause Redis to store the float as a string instead of a comma.

This change will make sure that before a float is stored it's always made sure that any comma is changed to a point. This was already done in some other parts of the system.

Fixes https://github.com/laravel/horizon/issues/371