Trim job IDs from failed job tags
Everytime a job fails, the ID is added to each tag. If there is no other failure in the next 48 hours, the tags expires.  But otherwise, the `StoreTagsForFailedJob` keeps adding job IDs and updates the expiration date, but the previous IDs were **never** removed, which probably causes memory issues.

In this commit we'll remove the job IDs for that are older than 48 hours for each tag, in respect of the TTL in `StoreTagsForFailedJob`.